+++ CHANGELOG.md
@@ -1,4 +1,24 @@
-# [0.4.1] - Minor improvements & New Extensions
+# [0.4.2] - Minor improvements & New Extensions
+
+- Returns the `String` after a character / word.
+
+```dart
+after(String pattern)
+```
+
+- Returns the `String` before a character / word.
+
+```dart
+before(String pattern)
+```
+
+- Gets Jaro distance to check `String` similarity. The higher the number the similar the two `String`s/
+
+```dart
+getJaro(String wordToCompare)
+```
+
+## [0.4.1] - Minor improvements & New Extensions
 
 - Converts the `String` to a `bool` value if possible.